+++ gateway/src/apicast/policy/local_chain/init.lua
@@ -1,0 +1,1 @@
+return require('local_chain')